+++ environment.yml
@@ -1,0 +1,18 @@
+name: mi-go
+channels:
+  - pytorch
+  - conda-forge
+  - defaults
+dependencies:
+  - cudatoolkit=12.1
+  - python=3.10
+  - pytorch
+  - youtube-dl
+  - ffmpeg
+  - pip:
+    - loguru
+    - pytube
+    - jiwer
+    - google-api-python-client
+    - youtube-transcript-api
+    - yt-dlp
