+++ libs/youtube_video.py
@@ -48,6 +48,7 @@ class YouTubeVideo:
 
             ydl_opts = {
                 "format": "bestaudio/best",
+                ""
                 "postprocessors": [
                     {
                         "key": "FFmpegExtractAudio",
