+++ models/__init__.py
@@ -1,0 +1,2 @@
+from models.dummy_test import DummyTest
+from models.whisper_test import WhisperTest
