+++ models/whisper_test.py
@@ -8,40 +8,7 @@ from whisper.normalizers import EnglishTextNormalizer
 
 import databases
 from libs.differs import jiwer_differ
-from testrunners.tests import TranscriptTest
-
-
-class DummyTest(TranscriptTest):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-        self.language = "en"
-        self.transcriber = lambda x: "This is a model for tests :)"
-        self.normalizer = lambda x: x.lower()
-        self.differ = jiwer_differ
-
-    def transcribe(self, audio_path: Path) -> str:
-        return self.transcriber(audio_path)
-
-    def compare(self, model_transcript, target_transcript):
-        normalized_model = self.normalizer(model_transcript)
-        normalized_target = self.normalizer(target_transcript)
-
-        differ_results = jiwer_differ(normalized_model, normalized_target)
-        differ_results["detectedLanguage"] = self.language
-
-        return differ_results
-
-    def testplan_postprocess(self, testplan):
-        testplan["model"] = {
-            "name": "Dummy",
-            "language": "en",
-            "werMean": 0.0,
-            "werStd": 1.0,
-        }
-
-    def insert_to_database(self, testplan):
-        databases.insert_transcript_diff_results(testplan, self.differ)
+from testrunners.transcript_test import TranscriptTest
 
 
 class WhisperTest(TranscriptTest):
@@ -50,16 +17,11 @@ class WhisperTest(TranscriptTest):
     """
 
     def __init__(
-        self,
-        model_type: str,
-        model_language: str = None,
-        gpu: int = 0,
-        **kwargs,
+        self, model_type: str, model_language: str = None, gpu: int = 0, **kwargs
     ):
         super().__init__(**kwargs)
         self.model_type = model_type
         self.model_language = model_language
-
         self.model = whisper.load_model(model_type, device=torch.device(f"cuda:{gpu}"))
 
         self.normalizer = EnglishTextNormalizer()
