+++ testrunners/youtube_runner.py
@@ -12,10 +12,11 @@ from loguru import logger
 from generators.youtube_generator import generate
 from libs.youtube_video import YouTubeVideo
 from testrunners import TestRegistry, TestRunner
-from testrunners.tests.tests import DummyTest, WhisperTest
+from models import DummyTest, WhisperTest
+from models.nemo_test import NemoTest
 
 
-@TestRegistry.register(WhisperTest, DummyTest)
+@TestRegistry.register(WhisperTest, DummyTest, NemoTest)
 class YouTubeTestRunner(TestRunner):
     """
     Test runner for youtube videos, uses a testplan generated by youtube_generator
